// src/modules/user/api/user.schema.ts
import { z } from 'zod';

export type UserStorageSchema = z.infer<typeof userStorageSchema>;

export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
}

export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: Geo;
}

export interface Geo {
  lat: string;
  lng: string;
}

export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

export const userStorageSchema = z.object({
  email: z.string(),
  token: z.string(),
});